\name{Alelager}
\alias{Alelager}
\docType{data}
\title{
Alcohol content and calories of beers
}
\description{
Alcohol content and calories for a sample of ale and lager beers.
}
\usage{Alelager}
\format{
  A data frame with 31 observations on the following 4 variables.
  \describe{
    \item{\code{ID}}{Subject ID}
    \item{\code{Type}}{Beer: ale or lager}
    \item{\code{Alcohol}}{Percentage alcohol content}
    \item{\code{Calories}}{Number of calories}
  }
}
%\examples{
%t.test(Calories ~ Type, Alelager, alt = "greater")
%}
\keyword{datasets}
