\name{Bangladesh}
\alias{Bangladesh}
\docType{data}
\title{
Arsenic levels of wells in Bangladesh
}
\description{
Levels of arsenic, chlorine and cobalt in a sample of 271 wells in Bangladesh.
}
\usage{Bangladesh}
\format{
  A data frame with 271 observations on the following 3 variables.
  \describe{
    \item{\code{Arsenic}}{Arsenic level, ppb}
    \item{\code{Chlorine}}{Chlorine level, ppb}
    \item{\code{Cobalt}}{Cobalt level, ppb}
  }
}
\source{
https://www2.bgs.ac.uk/groundwater/health/arsenic/Bangladesh/data.html
}
\references{
Reproduced with the permission of the British Geological Survey, copyright UKRI. All Rights Reserved.
}
%\examples{
%}
\keyword{datasets}
