\name{Beerwings}
\alias{Beerwings}
\docType{data}
\title{
Beer and hotwings consumption
}
\description{
Beer and hotwings consumption by a sample of patrons at a Minneapolis bar.
}
\usage{Beerwings}
\format{
  A data frame with 30 observations on the following 4 variables.
  \describe{
    \item{\code{ID}}{Subject ID}
    \item{\code{Hotwings}}{Number of hotwings consumed}
    \item{\code{Beer}}{Ounces of beers consumed}
    \item{\code{Gender}}{Gender of patron (M/F)}
  }
}
\source{
Data collected by Nicole Catchpole in 2004 (private communication).
}
%\examples{
%}

\keyword{datasets}
