\name{BookPrices}
\alias{BookPrices}
\docType{data}
\title{
Price of textooks at a college bookstore
}
\description{
Price of textbooks at a college bookstore.
}
\usage{BookPrices}
\format{
  A data frame with 44 observations on the following 3 variables.
  \describe{
    \item{\code{Subject}}{\code{Biology} \code{Chemistry} \code{Computer Science} \code{Economics} \code{Educational Studies} \code{Geology} \code{Mathematics} \code{Physics} \code{Political Science} \code{Psychology} \code{SOAN}}
    \item{\code{Area}}{Classification of subject as either \code{Math & Science} or \code{Social Sciences}}
    \item{\code{Price}}{Price in U.S.~dollars}
  }
}

\source{
Data collected by R.~Hien and S.~Becker in 2010 (private communication).
}
\keyword{datasets}
