\name{Bushmeat}
\alias{Bushmeat}
\docType{data}
\title{
Fish supply and demand for bushmeat in Ghana
}
\description{
Fish supply (kg) and demand for bushmeat in Ghana.
}
\usage{Bushmeat}
\format{
  A data frame with 30 observations on the following 4 variables.
  \describe{
    \item{\code{Fish}}{Fish supply (in kg.) per capita}
    \item{\code{Biomass}}{Biomass}
    \item{\code{Year}}{Year}
    \item{\code{Change}}{Percent change in biomass}
  }
}
\details{
Biomass of large mammals was calculated for each year by multiplying the
number of animals observed in 700 walking counts of 10 to 15 km each by
species-specific body weights. The products of these calculations were
then summed across all species.
}
\source{
Brashares, Arces, Sam, Coppolillo, Sinclaire, Balmford, Bushmeat hunting,
wildlife declines, and fish supply in West Africa, Science. 2004 Nov 12.
}
\keyword{datasets}
