\name{Cafeteria}
\alias{Cafeteria}
\docType{data}
\title{
Cafeteria
}
\description{
Nutritional data on meals served in a college cafeteria.
}
\usage{Cafeteria}
\format{
  A data frame with 41 observations on the following 9 variables.
  \describe{
    \item{\code{ID}}{a numeric vector}
    \item{\code{Type}}{type of meal, \code{Meat} or \code{Vegetarian}}
    \item{\code{Calories}}{number of calories}
    \item{\code{Carbohydrates}}{number of carbohydrates}
    \item{\code{Fiber}}{fiber content}
    \item{\code{Fat}}{fat content}
    \item{\code{Cholesterol}}{cholesterol}
    \item{\code{Protein}}{protein}
    \item{\code{Sodium}}{sodium}
  }
}

\source{
Stephenson (private communication).
}
\keyword{datasets}
