\name{Cereals}
\alias{Cereals}
\docType{data}
\title{
Cereals
}
\description{
Nutritional data on a sample of cereals.
}
\usage{Cereals}
\format{
  A data frame with 43 observations on the following 5 variables.
  \describe{
    \item{\code{ID}}{a numeric vector}
    \item{\code{Age}}{target consumer, \code{adult} or \code{children}}
    \item{\code{Shelf}}{location of cereal, \code{bottom}, \code{middle}, or \code{top} shelf}
    \item{\code{Sodiumgram}}{sodium content in grams}
    \item{\code{Proteingram}}{protein content in grams}
  }
}

\keyword{datasets}
