\name{Challenger}
\alias{Challenger}
\docType{data}
\title{
Challenger
}
\description{
Data on O-rings in 23 space shuttle flights prior to the Challenger shuttle disaster of January 1986.
}
\usage{Challenger}
\format{
  A data frame with 23 observations on the following 3 variables.
  \describe{
    \item{\code{Date}}{Data of launch}
    \item{\code{Temperature}}{Air temperature at launch (F)}
    \item{\code{Incident}}{Binary variable, 1 if one of the 0-rings on one of the booster rockets was damaged, 0 otherwise}
  }
}
\source{
https://archive.ics.uci.edu/ml/datasets/Challenger+USA+Space+Shuttle+O-Ring
}
\references{
Dala, S.~R., Fowlkes, E.~B., Hoadley, B (1989). Risk analysis of the space shuttle:
pre-Challenger prediction of failure. J.~American Statistical Association, 84, 945-957.
}

\keyword{datasets}
