\name{ChiMarathonMen}
\alias{ChiMarathonMen}
\docType{data}
\title{
ChiMarathonMen
}
\description{
Times from a sample of men who completed the Chicago marathon in 2015.
}
\usage{data("ChiMarathonMen")}
\format{
  A data frame with 80 observations on the following 4 variables.
  \describe{
    \item{\code{name}}{Name of competitor}
    \item{\code{Division}}{Age group}
    \item{\code{Finish}}{Finish time}
    \item{\code{FinishMin}}{Time in minutes}
  }
}
\source{
https://chicago-history.r.mikatiming.com/
}
\keyword{datasets}
