\name{Cuckoos}
\alias{Cuckoos}
\docType{data}
\title{
Cuckoos
}
\description{
Female cuckoos lay their eggs on the ground and then move them to the nests of other birds. Latter gathered data on the
lengths of the cuckoo eggs found in these foster nests.
}
\usage{data("Cuckoos")}
\format{
  A data frame with 120 observations on the following 2 variables.
  \describe{
    \item{\code{Eggs}}{Lengths of eggs (mm) of cuckoos}
    \item{\code{Bird}}{Species of birds: \code{HedgeSparrow}, \code{MeadowPipit}, \code{PiedWagtail}, \code{Robin}, \code{TreePipit}, \code{Wren}}
  }
}
\source{
Tippett, L. H. C. (1952). The Methods of Statistics, 4th Edition. Wiley. 
}
\references{
Latter, O. (1902). An enquiry into the dimensions of the Cuckoo's egg and the relation of the variations to the size
of eggs of the foster-parent, with notes on coloration. Biometrika 1 (2): 164-176.
}

\keyword{datasets}
