\name{Diving2017}
\alias{Diving2017}
\docType{data}
\title{
Diving 2017
}
\description{
Scores of 12 female divers (10 m platform) in the 2017 FINA World Championships.
}
\usage{data("Diving2017")}
\format{
  A data frame with 12 observations on the following 4 variables.
  \describe{
    \item{\code{Name}}{Name of competitor}
    \item{\code{Country}}{Country}
    \item{\code{Semifinal}}{Score in the semi-finals}
    \item{\code{Final}}{Score in the finals.}
  }
}
\details{
Competitors perform 5 dives in each round and the sum of these 5 dives determines who moves on to the next round.
}
\source{
https://www.fina.org/competitions/213/17th-fina-world-championships-2017/results?disciplines=DV
}
\keyword{datasets}
