\name{Eyes}
\alias{Eyes}
\docType{data}
\title{
Eyes
}
\description{
Measurements of eyes of 40 people.
}
\usage{data("Eyes")}
\format{
  A data frame with 40 observations on the following 6 variables.
  \describe{
    \item{\code{ID}}{Subject ID}
    \item{\code{age}}{Age of subject}
    \item{\code{hand}}{Dominant hand of subject, \code{left} or \code{right}}
    \item{\code{eye}}{Dominant eye of subject, \code{left} or \code{right}}
    \item{\code{leftPD}}{Left pupillary distance (mm)}
    \item{\code{rightPD}}{Right pupillary distance (mm)}
  }
}
\source{
Westfield (private communication).
}
\keyword{datasets}
