\name{Fatalities}
\alias{Fatalities}
\docType{data}
\title{
Fatalities
}
\description{
A random sample of driver fatalities in 2009 in Pennsylvania.
}
\usage{Fatalities}
\format{
  A data frame with 100 observations on the following 3 variables.
  \describe{
    \item{\code{ID}}{Subject ID}
    \item{\code{Alcohol}}{Alcohol involved? 1 = yes, 0 = no}
    \item{\code{Age}}{Age}
  }
}
\details{
The drivers were driving a car, SUV, or light pickup truck (vehicles such as motor homes, convertibles, or
commercial vehicles are excluded).
}
\source{
http://www.nhtsa.gov/FARS
}

\keyword{datasets}
