\name{FishMercury}
\alias{FishMercury}
\docType{data}
\title{
Mercury content in a sample of fish in Minnesota
}
\description{
Mercury levels (ppm) in a sample of fish caught in Minnesota
}
\usage{FishMercury}
\format{
  A data frame with 30 observations on the following variable.
  \describe{
    \item{\code{Mercury}}{Mercury level in ppm}
  }
}
\source{
Minnesota pollution control agency.
}

\keyword{datasets}
