\name{Girls2004}
\alias{Girls2004}
\docType{data}
\title{
Births of girls in Alaska or Wyoming
}
\description{
Data on births of a random sample of girls in Alaska or Wyoming in 2004.
}
\usage{data("Girls2004")}
\format{
  A data frame with 80 observations on the following 6 variables.
  \describe{
    \item{\code{ID}}{Subject ID}
    \item{\code{State}}{State: \code{AK} or \code{WY}}
    \item{\code{MothersAge}}{Age of mother: \code{15-19}, \code{20-24}, \code{25-29}, \code{30-34}, \code{35-39}, \code{40-44}}
    \item{\code{Smoker}}{Mother a smoker? \code{No} or \code{Yes}}
    \item{\code{Weight}}{Weight of baby (grams)}
    \item{\code{Gestation}}{Gestation time (weeks)}
  }
}
\source{
http://wonder.cdc.gov/natality-current.html
}

\keyword{datasets}
