\name{Groceries}
\alias{Groceries}
\docType{data}
\title{
Groceries
}
\description{
Prices of a sample of grocery items at Target or Walmart.
}
\usage{Groceries}
\format{
  A data frame with 30 observations on the following 4 variables.
  \describe{
    \item{\code{Product}}{Grocery item}
    \item{\code{Size}}{Package size}
    \item{\code{Target}}{Price at Target}
    \item{\code{Walmart}}{Price at Walmart}
  }
}

\keyword{datasets}
