\name{GSS2018}
\alias{GSS2018}
\docType{data}
\title{
General Social Survey 2018
}
\description{
General Social Survey data from 2018
}
\usage{GSS2018}
\format{
  A data frame with 2348 observations on the following 17 variables.
  \describe{
    \item{\code{ID}}{Subject ID}
     \item{\code{Region}}{\code{Midwest}, \code{Northeast}, \code{South}, \code{West}}
     \item{\code{GenderNow}}{Gender of subject: \code{A gender not listed here}, \code{Man}, \code{Not applicable}, \code{Transgender}, \code{Woman}}
    \item{\code{Age}}{Age}
    \item{\code{Marital}}{Marital status: \code{Divorced}, \code{Married}, \code{Never married}, \code{Separated}, \code{Widowed}}
    \item{\code{Degree}}{Education: \code{Bachelor} \code{Graduate}, \code{High school} \code{Junior college}, \code{Less than high school}}
    \item{\code{Employed}}{Employed? \code{No} or \code{Yes}}
    \item{\code{Income}}{Income level}
    \item{\code{Polviews}}{Political views:\code{Conservative}, \code{Extremely liberal}, \code{Extremely conservative}, \code{Liberal}, \code{Moderate}, \code{Slightly conservative}, \code{Slightly liberal}}
     \item{\code{Pres16}}{Voted for whom in presidential election of 2016? \code{Clinton}, \code{Other}, \code{Trump}}
 \item{\code{DeathPenalty}}{Opinion on death penalty: \code{Favor}, \code{Oppose}}
  \item{\code{Courts}}{How courts deal with criminals: \code{About right}, \code{Dont know}, \code{Not harsh enough}, \code{Too harsh}}
 \item{\code{Attend}}{Attendance at religious services: \code{Monthly}, \code{Never}, \code{Occasionally}, \code{Weekly}}
 \item{\code{Postlife}}{Believe in life after death? \code{Dont know}, \code{No}, \code{Yes}}
    \item{\code{Happy}}{General happiness level: \code{Not too happy}, \code{Pretty happy}, \code{Very happy}}
    \item{\code{Satfin}}{Satisfaction with financial situation: \code{More or less}, \code{Not at all}, \code{Satisfied}}
    \item{\code{Energy}}{Government spending on developing alternative energy sources: \code{About right}, \code{Dont know}, \code{Too little}, \code{Too much}}
           }
}
\source{
https://gss.norc.org
}
\keyword{datasets}
