\name{IceCream}
\alias{IceCream}
\docType{data}
\title{
Ice Cream
}
\description{
Nutritional information on a sample of ice cream.
}
\usage{data("IceCream")}
\format{
  A data frame with 39 observations on the following 7 variables.
  \describe{
    \item{\code{Brand}}{Brand of ice cream}
    \item{\code{VanillaCalories}}{Calories in vanilla}
    \item{\code{VanillaFat}}{Fat (gm) in vanilla ice cream}
    \item{\code{VanillaSugar}}{Sugar (gm) in vanilla ice cream}
    \item{\code{ChocolateCalories}}{Calories in chocolate ice cream}
    \item{\code{ChocolateFat}}{Fat (gm) in chocolate ice cream}
    \item{\code{ChocolateSugar}}{Sugar (gm) in chocolate ice cream}
  }
}
\keyword{datasets}
