\name{ILBoys}
\alias{ILBoys}
\docType{data}
\title{
Birth weight of boys born in Illinois
}
\description{
Birth weight of boys born in Illinois.
}
\usage{ILBoys}
\format{
  A data frame with 241 observations on the following 2 variables.
  \describe{
    \item{\code{MothersAge}}{Age range of mother: \code{15-19}, \code{20-24}, \code{25-29}}
    \item{\code{Weight}}{Weight of baby (gm)}
  }
}
\details{
Random sample of boys born to mothers in Illinois in 2004. Births are restricted to single births only and
gestation lengths of at least 37 weeks.
}

\keyword{datasets}
