\name{Illiteracy}
\alias{Illiteracy}
\docType{data}
\title{
Illiteracy
}
\description{
Data on female illiteracy in a sample of countries where illiteracy is more than 5\%.
}
\usage{Illiteracy}
\format{
  A data frame with 94 observations on the following 4 variables.
  \describe{
    \item{\code{ID}}{Country ID}
    \item{\code{Country}}{Name of country}
    \item{\code{Illit}}{Percentage of women over 15 years old who are illiterate (2003)}
    \item{\code{Births}}{Number of births per woman in that country (2005)}
  }
}



\source{
www.unesco.org, www.data.worldbank.org
}


\keyword{datasets}
