\name{Lottery}
\alias{Lottery}
\docType{data}
\title{
Lottery
}
\description{
Winning lottery numbers for Fantasy 5 in California.
}
\usage{Lottery}
\format{
  A data frame with 500 observations on the following variable.
  \describe{
    \item{\code{Win}}{Number}
  }
}
\details{
In Fantasy 5, a lottery game in California, a player tries to match 5 numbers
chosen from 1 through 39. This data are the winning numbers for the daily games
from 5 May 2010 through 15 August 2010.
}
\source{
http://www.calottery.com/play/draw-games/fantasy-5
}

\keyword{datasets}
