\name{MathAnxiety}
\alias{MathAnxiety}
\docType{data}
\title{
Math Anxiety
}
\description{
Data from a study on math anxiety in a sample of primary and secondary school students in
Italy
}
\usage{MathAnxiety}
\format{
  A data frame with 599 observations on the following 6 variables.
  \describe{
    \item{\code{Age}}{Age}
    \item{\code{Gender}}{Gender: \code{Boy}, \code{Girl}}
    \item{\code{Grade}}{Grade: \code{Secondary}, \code{Primary}}
    \item{\code{AMAS}}{Score on Abbreviated Math Anxiety Scale}
    \item{\code{RCMAS}}{Score on Revised Abbreviated Math Anxiety Scale}
    \item{\code{Arith}}{Score on arithmetic test }
  }
}
\source{
Hill, Mammarella, Devine, et al (2016). Maths anxiety in primary and secondary school
students: gender differences, developmental changes and anxiety specificity.
Learning and Individual Differences 48: 45-53
}

\keyword{datasets}
