\name{Maunaloa}
\alias{Maunaloa}
\docType{data}
\title{
Carbon dioxide levels collected by Mauna Loa Observatory
}
\description{
Average CO2 levels (ppm) for the month of May from 1990 to 2010.
}
\usage{Maunaloa}
\format{
  A data frame with 21 observations on the following 3 variables.
  \describe{
    \item{\code{ID}}{Subject ID}
    \item{\code{Year}}{Year}
    \item{\code{Level}}{Carbon dioxide level (ppm)}
  }
}
\source{
https://www.esrl.noaa.gov/gmd/ccgg/trends
}
\keyword{datasets}
