\name{MnGroundwater}
\alias{MnGroundwater}
\docType{data}
\title{
Minnesota groundwater
}
\description{
Measurements on water quality in wells in Minnesota.
}
\usage{MnGroundwater}
\format{
  A data frame with 895 observations on the following 10 variables.
  \describe{
    \item{\code{County}}{Minnesota county}
    \item{\code{Aquifer.Group}}{Type of aquifer: \code{buried Quaternary}, \code{Cambrian}, \code{Cretaceous}, \code{Devonian}, \code{Ordovician}, \code{Precambrian}, \code{surficial Quaternary}}
    \item{\code{Water.Level}}{Water level}
    \item{\code{Alkalinity}}{Alkalinity}
    \item{\code{Aluminum}}{Aluminum}
    \item{\code{Arsenic}}{Arsenic}
    \item{\code{Chloride}}{Chloride}
    \item{\code{Lead}}{lead}
    \item{\code{pH}}{pH level}
    \item{\code{Basin.Name}}{Basin name}
  }
}
\source{
Minnesota Pollution Control Agency
}
\keyword{datasets}
