\name{Nasdaq}
\alias{Nasdaq}
\docType{data}
\title{
Nasdaq stock data
}
\description{
Opening and closing stock prices for a random sample of 50 stock funds on NASDAQ on 1 December 2017.
}
\usage{Nasdaq}
\format{
  A data frame with 50 observations on the following 4 variables.
  \describe{
    \item{\code{Symbol}}{Stock symbol}
    \item{\code{Open}}{Opening price}
    \item{\code{Close}}{Closing price}
    \item{\code{Volume}}{Number of shares traded}
  }
}

\source{
https://finance.yahoo.com
}


\keyword{datasets}
