\name{Olympics2012}
\alias{Olympics2012}
\docType{data}
\title{
Olympics 2012
}
\description{
Data on a sample of athletes competing in the 2012 London Olympics.
}
\usage{Olympics2012}
\format{
  A data frame with 42 observations on the following 7 variables.
  \describe{
    \item{\code{Name}}{Name of athlete}
    \item{\code{Country}}{Country}
    \item{\code{Age}}{Age}
    \item{\code{Sex}}{Sex: \code{F}, \code{M}}
    \item{\code{Height}}{Height (inches)}
    \item{\code{Weight}}{weight (lb)}
    \item{\code{Sport}}{Sport}
  }
}


\keyword{datasets}
