\name{Oscars}
\alias{Oscars}
\docType{data}
\title{
Oscars
}
\description{
Age and gender of Academy Award winners
}
\usage{Oscars}
\format{
  A data frame with 188 observations on the following 6 variables.
  \describe{
    \item{\code{Year}}{Year of award}
    \item{\code{Actor}}{Name of actor}
    \item{\code{Movie}}{Movie}
    \item{\code{Gender}}{Gender: \code{Man}, \code{Woman}}
    \item{\code{Birthyear}}{Birth year of actor}
    \item{\code{Age}}{Age at time of award}
  }
}

\source{
https://www.oscars.org/
}

\keyword{datasets}
