\name{Phillies2009}
\alias{Phillies2009}
\docType{data}
\title{
Philadelphia Phillies data 2009
}
\description{
Baseball data for Philadelphia Phillies during the 2009 season.
}
\usage{Phillies2009}
\format{
  A data frame with 162 observations on the following 8 variables.
  \describe{
    \item{\code{Date}}{Date of game}
    \item{\code{Location}}{Game played where: \code{Away}, \code{Home}}
    \item{\code{Outcome}}{Outcome of game: \code{Lose}, \code{Win}}
    \item{\code{Outcome2}}{Outcome recoded: 1=win, 0 = lose}
    \item{\code{Hits}}{Number of hits}
    \item{\code{Doubles}}{Number of doubles}
    \item{\code{Homeruns}}{Number of homeruns}
    \item{\code{StrikeOuts}}{Number of strikeouts}
  }
}

\source{
https://www.baseball-reference.com/
}

\keyword{datasets}
