\name{Quakes}
\alias{Quakes}
\docType{data}
\title{
Quakes
}
\description{
Time between earthquakes for all earthquakes of magnitude 6 or greater (1970-2009).
}
\usage{data("Quakes")}
\format{
  A data frame with 805 observations on the following 2 variables.
  \describe{
    \item{\code{ID}}{Subject ID}
    \item{\code{TimeDiff}}{Time (days)}
  }
}

\source{
http://earthquakes.usgs.gov/earthquakes/eqarchives
}

\keyword{datasets}
