\name{Quetzal}
\alias{Quetzal}
\docType{data}
\title{
Quetzal
}
\description{
Heights of nests and snags for the quetzal (bird).
}
\usage{Quetzal}
\format{
  A data frame with 21 observations on the following 3 variables.
  \describe{
    \item{\code{Country}}{Country: \code{Costa Rica}, \code{Guatemala}}
    \item{\code{Nest}}{Height of nest (meters)}
    \item{\code{Snag}}{Height of snag (meters)}
  }
}
\details{
The quetzal typically nests in abandoned woodpecker nests in dead tree trunks (snags).
}
\source{
Siegfried, D., Linville, D., Hille, D. (2010). Analysis of nest sites and the resplendent quetzal (pharomachrus mocinno):
relationship between nest and snag heights. Wilson Journal of Ornithology 122: 608-11.
}
\keyword{datasets}
