\name{RangersTwins2016}
\alias{RangersTwins2016}
\docType{data}
\title{
Rangers and Twins baseball players (2016 season)
}
\description{
Data on baseball players (excluding pitchers) who played for the Texas Rangers or Minnesota Twins.
}
\usage{data("RangersTwins2016")}
\format{
  A data frame with 27 observations on the following 17 variables.
  \describe{
    \item{\code{Name}}{Name of player}
    \item{\code{Team}}{Team: \code{Rangers}, \code{Twins}}
    \item{\code{Pos}}{Player's position}
    \item{\code{Age}}{Age in years}
    \item{\code{Games}}{Number of games played}
    \item{\code{AtBats}}{Number of at bats}
    \item{\code{Runs}}{Runs}
    \item{\code{Hits}}{Hits}
    \item{\code{Doubles}}{Doubles}
    \item{\code{Triples}}{Triples}
    \item{\code{HR}}{Homeruns}
    \item{\code{RBI}}{Runs batted in}
    \item{\code{SB}}{Stolen bases}
    \item{\code{CS}}{Caught stealing}
    \item{\code{BB}}{Base on balls}
    \item{\code{SO}}{Strike outs}
    \item{\code{BA}}{Batting average}
  }
}
\details{
Data on baseball players (excluding pitchers) who played for the Texas Rangers or Minnesota Twins.
These players played at least 50 games. During the 2016 season, the Rangers had the best winning percentage (0.586) in the
American League while the Twins had the worst (0.364)} 
\source{
www.baseball-reference.com
}
\keyword{datasets}
