\name{resampledata3-package}
\alias{resampledata3-package}
\alias{resampledata3}
\docType{package}
\title{
\packageTitle{resampledata3}
}
\description{
\packageDescription{resampledata3}
\url{https://github.com/lchihara/MathStatsResamplingR}
}
\examples{
# For a list of datasets do:
library(help = resampledata3)
}
