\name{Salaries}
\alias{Salaries}
\docType{data}
\title{
Salaries of baseball players
}
\description{
Salaries of a random sample of baseball players from 1985 and 2015.
}
\usage{Salaries}
\format{
  A data frame with 70 observations on the following 3 variables.
  \describe{
    \item{\code{League}}{League: \code{American} \code{National}}
    \item{\code{Salary}}{Salary (in millions) in 2015 dollars}
    \item{\code{Year}}{Year: 1985 or 2015}
  }
}

\keyword{datasets}
