\name{Service}
\alias{Service}
\docType{data}
\title{
Service times at a college snack bar.
}
\description{
Time to be served at a college snack bar.
}
\usage{Service}
\format{
  A data frame with 174 observations on the following 2 variables.
  \describe{
    \item{\code{ID}}{Subject ID}
    \item{\code{Times}}{Time in minutes}
  }
}
\source{
Haynor, Lojovich, Syed (private communication, 2010).
}
\keyword{datasets}
