\name{Skateboard}
\alias{Skateboard}
\docType{data}
\title{
Skateboard experiment
}
\description{
Measurement of testosterone levels in males in a skateboard experiment.
}
\usage{Skateboard}
\format{
  A data frame with 71 observations on the following 3 variables.
  \describe{
    \item{\code{Age}}{Age in years}
    \item{\code{Experimenter}}{Treatment (gender of experimenter): \code{Female}, \code{Male}}
    \item{\code{Testosterone}}{Testosterone level}
  }
}
\details{
Results from an experiment where male skateboarders performed tricks in front of either 
a female or male.
}
\source{
Ronay and Hippel (2010). The presence of an attractive woman elevates testosterone and
physical risk taking in young men. Social Psychological and Personality Science 1:57-64.
}


\keyword{datasets}
