\name{Skating2010}
\alias{Skating2010}
\docType{data}
\title{
Figure skating scores for men from the 2010 Winter Olympics.
}
\description{
Short and free skate scores for male figure skaters in the 2010 Winter Olympics (Vancouver).
}
\usage{Skating2010}
\format{
  A data frame with 24 observations on the following 5 variables.
  \describe{
  \item{\code{Country}}{Country of skater}
    \item{\code{Name}}{Name}
    \item{\code{Short}}{Short program score}
    \item{\code{Free}}{Free skate score}
    \item{\code{Total}}{Total}
  }
}

\source{
https://skatingscores.com/0910/oly/
}

\keyword{datasets}
