\name{Spruce}
\alias{Spruce}
\docType{data}
\title{
Spruce data
}
\description{
Measurements from an experiment on the growth of black spruce seedlings.
}
\usage{Spruce}
\format{
  A data frame with 72 observations on the following 9 variables.
  \describe{
    \item{\code{Tree}}{Subject ID}
    \item{\code{Competition}}{Treatment: \code{C} (competition), \code{NC} (no competition)}
    \item{\code{Fertilizer}}{Treatment: \code{F} (fertilizer), \code{NF} (no fertilizer)}
    \item{\code{Height0}}{Height of seedling at start}
    \item{\code{Height5}}{Height of seedling after 5 years}
    \item{\code{Diameter0}}{Diameter of seedling at start}
    \item{\code{Diameter5}}{Diameter of seedling after 5 years}
    \item{\code{Ht.change}}{Change in height}
    \item{\code{Di.change}}{Change in diameter}
  }
}
\details{
Experiment on growth of black spruce seedlings under treatments of fertilizer-no fertilizer, competition-
no competition (weeding).
}
\source{
Camill, Chihara, Adams, et al (2010). Early life history transitions and recruitment of Picea mariana in
thawed boreal permafrost peatlands. Ecology 2:448-459.
}

\keyword{datasets}
