\name{Starcraft}
\alias{Starcraft}
\docType{data}
\title{
Starcraft
}
\description{
Number of wins by a sample of Korean players in Starcraft, a strategy video game.
}
\usage{Starcraft}
\format{
  A data frame with 45 observations on the following 4 variables.
  \describe{
    \item{\code{ID}}{Subject ID}
    \item{\code{Race}}{Chosen race of player: \code{Protoss}, \code{Terran}, \code{Zerg}}
    \item{\code{Age}}{Age of player}
    \item{\code{Wins}}{Number of wins}
  }
}
\source{
Evans, private communication. http://www.teamliquid.net/tipd/players 
}

\keyword{datasets}
