\name{Titanic}
\alias{Titanic}
\docType{data}
\title{
Titanic
}
\description{
Subset of Titanic data
}
\usage{Titanic}
\format{
  A data frame with 658 observations on the following 3 variables.
  \describe{
    \item{\code{ID}}{Subject ID}
    \item{\code{Survived}}{Survival status: 1 = survived, 0 = died}
    \item{\code{Age}}{Age of passenger}
  }
}
\details{
Subset of passenger data on the Titanic.
}
\source{
https://data.world/nrippner/titanic-disaster-dataset
}



\keyword{datasets}
