\name{Turbine}
\alias{Turbine}
\docType{data}
\title{
Turbine
}
\description{
Average daily wind speeds (2010) from Carleton College turbine.
}
\usage{data("Turbine")}
\format{
  A data frame with 168 observations on the following 4 variables.
  \describe{
    \item{\code{Date2010}}{Date}
    \item{\code{AveKW}}{Average kilowatts}
    \item{\code{AveSpeed}}{Average speed (m/s)}
    \item{\code{Production}}{Energy output (kilowatt hours)}
  }
}

\source{
Carleton College, Northfield MN.
}
\references{
Chihara and Hesterberg (2022). Mathematical Statistics with Resampling and R. (Wiley)
}


\keyword{datasets}
