\name{TV}
\alias{TV}
\docType{data}
\title{
TV commercial lengths
}
\description{
Lengths of television commercials on basic and extended cable TV channel.s
}
\usage{data("TV")}
\format{
  A data frame with 20 observations on the following 3 variables.
  \describe{
    \item{\code{ID}}{Subject ID}
    \item{\code{Times}}{Time (min)}
    \item{\code{Cable}}{Cable: \code{Basic}, \code{Extended}}
  }
}
\details{
Lengths of TV commercials during any given half-hour time period.
}
\source{
Rodgers, Robinson (private communication).
}

\keyword{datasets}
