\name{TXBirths2004}
\alias{TXBirths2004}
\docType{data}
\title{
Texas birth weights 
}
\description{
Weights of babies born in Texas in 2004.
}
\usage{TXBirths2004}
\format{
  A data frame with 1587 observations on the following 8 variables.
  \describe{
    \item{\code{ID}}{Subject ID}
    \item{\code{MothersAge}}{Mother's age: \code{15-19}, \code{20-24}, \code{25-29}, \code{30-34}, \code{35-39}, \code{40-44}, \code{under 15}}
    \item{\code{Smoker}}{Mother smokes? \code{No}, \code{Yes}}
    \item{\code{Gender}}{Gender of baby: \code{Female}, \code{Male}}
    \item{\code{Weight}}{Weight of baby (g)}
    \item{\code{Gestation}}{Gestation length (weeks)}
    \item{\code{Number}}{Baby a single birth (1), twin (2), etc.}
    \item{\code{Multiple}}{Part of multiple birth (eg twin, triple)?: \code{No}, \code{Yes}}
  }
}
\details{
Random sample of babies born in Texas in 2004. 
}
\source{
http://wonder.cdc.gov/natality-current.html
}

\keyword{datasets}
