\name{Volleyball2009}
\alias{Volleyball2009}
\docType{data}
\title{
Volleyball data
}
\description{
Data on a sample of Division I women volleyball teams.
}
\usage{Volleyball2009}
\format{
  A data frame with 30 observations on the following 4 variables.
  \describe{
    \item{\code{Team}}{Team}
    \item{\code{HitPercent}}{Hitting percentage}
    \item{\code{Assts}}{Assists}
    \item{\code{Kills}}{Kills}
  }
}
\source{
http://www.ncaa.org/championships/statistics/womens-volleyball-statistics
}

\keyword{datasets}
