\name{Walleye}
\alias{Walleye}
\docType{data}
\title{
Walleye
}
\description{
Lengths and weights of a sample of walleye caught in Minnesota lakes (1990's).
}
\usage{Walleye}
\format{
  A data frame with 60 observations on the following 2 variables.
  \describe{
    \item{\code{Length}}{Length (inches)}
    \item{\code{Weight}}{Weight (pounds)}
  }
}

\source{
Monson, Minnesota Pollution Control Agency (private communication)
}

\keyword{datasets}
