\name{Watertable}
\alias{Watertable}
\docType{data}
\title{
Watertable
}
\description{
Relationship between the depth of the watertable and survival status of
black spruce seedings.
}
\usage{Watertable}
\format{
  A data frame with 360 observations on the following 2 variables.
  \describe{
    \item{\code{Depth}}{Depth of watertable (cm)}
    \item{\code{Alive}}{Status of seedling: 1 = alive, 0 = dead}
  }
}
\details{
Part of the data from an experiment to study factors associated with the growth of black spruce
seedlings under various treatments. Status of seedling at the end of the second year
of the experiment is noted here.
}
\source{
Camill, Chihara, Adams, et al (2010). Early life history transitions and recruitment of
Picae mariana in thawed boreal permafrost peatlands. Ecology 2:448-459.
}
\references{
Chihara and Hesterberg (2022). Mathematical Statistics with Resampling and R (Wiley).
}
\keyword{datasets}
